# backend/artifacts/openapi/devsphere-backend-final.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DevSphere.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/companies": {
      "post": {
        "tags": [
          "CompanyTrust"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyProfileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyProfileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/companies/{companyId}": {
      "put": {
        "tags": [
          "CompanyTrust"
        ],
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyProfileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyProfileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/companies/{companyId}/verification": {
      "post": {
        "tags": [
          "CompanyTrust"
        ],
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCompanyVerificationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCompanyVerificationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCompanyVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/profile/employer": {
      "get": {
        "tags": [
          "EmployerProfile"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "EmployerProfile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerProfileDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerProfileDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "EmployerProfile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerProfileDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerProfileDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/vacancies/{vacancyId}/policy/current": {
      "get": {
        "tags": [
          "VacancyPolicy"
        ],
        "parameters": [
          {
            "name": "vacancyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "VacancyPolicy"
        ],
        "parameters": [
          {
            "name": "vacancyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyPolicyAggregateUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyPolicyAggregateUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyPolicyAggregateUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vacancies/{vacancyId}/policy/current/full": {
      "get": {
        "tags": [
          "VacancyPolicy"
        ],
        "parameters": [
          {
            "name": "vacancyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyPolicyAggregateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyPolicyAggregateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyPolicyAggregateDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/vacancies/{vacancyId}/policy/families": {
      "post": {
        "tags": [
          "VacancyPolicy"
        ],
        "parameters": [
          {
            "name": "vacancyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyPolicyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyPolicyRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyPolicyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vacancies/{vacancyId}/policy/requirements": {
      "post": {
        "tags": [
          "VacancyPolicy"
        ],
        "parameters": [
          {
            "name": "vacancyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyRequirementRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyRequirementRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyRequirementRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vacancies/{vacancyId}/policy/alternative-sets": {
      "post": {
        "tags": [
          "VacancyPolicy"
        ],
        "parameters": [
          {
            "name": "vacancyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlternativeSetRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AlternativeSetRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AlternativeSetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AlternativeSetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "matchingPolicyRevisionId": {
            "type": "string",
            "format": "uuid"
          },
          "familyPolicyId": {
            "type": "string",
            "format": "uuid"
          },
          "setType": {
            "$ref": "#/components/schemas/AlternativeSetType"
          },
          "minimumSatisfiedCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "mode": {
            "$ref": "#/components/schemas/RequirementMode"
          },
          "importance": {
            "$ref": "#/components/schemas/RequirementImportance"
          },
          "isActive": {
            "type": "boolean"
          },
          "isScored": {
            "type": "boolean"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "memberRequirementIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AlternativeSetEditRequest": {
        "type": "object",
        "properties": {
          "familyPolicyId": {
            "type": "string",
            "format": "uuid"
          },
          "setType": {
            "$ref": "#/components/schemas/AlternativeSetType"
          },
          "minimumSatisfiedCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "mode": {
            "$ref": "#/components/schemas/RequirementMode"
          },
          "importance": {
            "$ref": "#/components/schemas/RequirementImportance"
          },
          "isActive": {
            "type": "boolean"
          },
          "isScored": {
            "type": "boolean"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "memberRequirementIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "clientKey": {
            "type": "string",
            "nullable": true
          },
          "familyClientKey": {
            "type": "string",
            "nullable": true
          },
          "memberRequirementClientKeys": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AlternativeSetRequest": {
        "type": "object",
        "properties": {
          "familyPolicyId": {
            "type": "string",
            "format": "uuid"
          },
          "setType": {
            "$ref": "#/components/schemas/AlternativeSetType"
          },
          "minimumSatisfiedCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "mode": {
            "$ref": "#/components/schemas/RequirementMode"
          },
          "importance": {
            "$ref": "#/components/schemas/RequirementImportance"
          },
          "isActive": {
            "type": "boolean"
          },
          "isScored": {
            "type": "boolean"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "memberRequirementIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateCompanyProfileRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EmployerProfileDto": {
        "type": "object",
        "properties": {
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "contactEmail": {
            "type": "string",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FamilyPolicyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "matchingPolicyRevisionId": {
            "type": "string",
            "format": "uuid"
          },
          "requirementFamily": {
            "$ref": "#/components/schemas/RequirementFamily"
          },
          "familyImportance": {
            "$ref": "#/components/schemas/RequirementImportance"
          },
          "isActive": {
            "type": "boolean"
          },
          "isScored": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "FamilyPolicyEditRequest": {
        "type": "object",
        "properties": {
          "requirementFamily": {
            "$ref": "#/components/schemas/RequirementFamily"
          },
          "familyImportance": {
            "$ref": "#/components/schemas/RequirementImportance"
          },
          "isActive": {
            "type": "boolean"
          },
          "isScored": {
            "type": "boolean"
          },
          "clientKey": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FamilyPolicyRequest": {
        "type": "object",
        "properties": {
          "requirementFamily": {
            "$ref": "#/components/schemas/RequirementFamily"
          },
          "familyImportance": {
            "$ref": "#/components/schemas/RequirementImportance"
          },
          "isActive": {
            "type": "boolean"
          },
          "isScored": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "MatchingPolicyRevisionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "vacancyId": {
            "type": "string",
            "format": "uuid"
          },
          "revisionNumber": {
            "type": "integer",
            "format": "int32"
          },
          "isCurrent": {
            "type": "boolean"
          },
          "isMateriallyLocked": {
            "type": "boolean"
          },
          "materiallyLockedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SubmitCompanyVerificationRequest": {
        "type": "object",
        "properties": {
          "evidenceStorageKey": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateCompanyProfileRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "VacancyPolicyAggregateDto": {
        "type": "object",
        "properties": {
          "revision": {
            "$ref": "#/components/schemas/MatchingPolicyRevisionDto"
          },
          "families": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FamilyPolicyDto"
            },
            "nullable": true
          },
          "requirements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VacancyRequirementPolicyDto"
            },
            "nullable": true
          },
          "alternativeSets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlternativeSetDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "VacancyPolicyAggregateUpdateRequest": {
        "type": "object",
        "properties": {
          "families": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FamilyPolicyEditRequest"
            },
            "nullable": true
          },
          "requirements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VacancyRequirementEditRequest"
            },
            "nullable": true
          },
          "alternativeSets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlternativeSetEditRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "VacancyRequirementEditRequest": {
        "type": "object",
        "properties": {
          "familyPolicyId": {
            "type": "string",
            "format": "uuid"
          },
          "requirementFamily": {
            "$ref": "#/components/schemas/RequirementFamily"
          },
          "mode": {
            "$ref": "#/components/schemas/RequirementMode"
          },
          "importance": {
            "$ref": "#/components/schemas/RequirementImportance"
          },
          "isActive": {
            "type": "boolean"
          },
          "isScored": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "skillConceptId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "canonicalTargetKey": {
            "type": "string",
            "nullable": true
          },
          "requiredMonths": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "requiredValue": {
            "type": "string",
            "nullable": true
          },
          "acceptedValuesJson": {
            "type": "string",
            "nullable": true
          },
          "isRegulatoryGate": {
            "type": "boolean"
          },
          "requiresVerification": {
            "type": "boolean"
          },
          "questionText": {
            "type": "string",
            "nullable": true
          },
          "expectedAnswer": {
            "type": "string",
            "nullable": true
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "clientKey": {
            "type": "string",
            "nullable": true
          },
          "familyClientKey": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "VacancyRequirementPolicyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "vacancyId": {
            "type": "string",
            "format": "uuid"
          },
          "matchingPolicyRevisionId": {
            "type": "string",
            "format": "uuid"
          },
          "familyPolicyId": {
            "type": "string",
            "format": "uuid"
          },
          "alternativeSetId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "requirementFamily": {
            "$ref": "#/components/schemas/RequirementFamily"
          },
          "mode": {
            "$ref": "#/components/schemas/RequirementMode"
          },
          "importance": {
            "$ref": "#/components/schemas/RequirementImportance"
          },
          "isActive": {
            "type": "boolean"
          },
          "isScored": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "skillConceptId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "canonicalTargetKey": {
            "type": "string",
            "nullable": true
          },
          "requiredMonths": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "requiredValue": {
            "type": "string",
            "nullable": true
          },
          "acceptedValuesJson": {
            "type": "string",
            "nullable": true
          },
          "isRegulatoryGate": {
            "type": "boolean"
          },
          "requiresVerification": {
            "type": "boolean"
          },
          "questionText": {
            "type": "string",
            "nullable": true
          },
          "expectedAnswer": {
            "type": "string",
            "nullable": true
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "VacancyRequirementRequest": {
        "type": "object",
        "properties": {
          "familyPolicyId": {
            "type": "string",
            "format": "uuid"
          },
          "requirementFamily": {
            "$ref": "#/components/schemas/RequirementFamily"
          },
          "mode": {
            "$ref": "#/components/schemas/RequirementMode"
          },
          "importance": {
            "$ref": "#/components/schemas/RequirementImportance"
          },
          "isActive": {
            "type": "boolean"
          },
          "isScored": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "skillConceptId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "canonicalTargetKey": {
            "type": "string",
            "nullable": true
          },
          "requiredMonths": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "requiredValue": {
            "type": "string",
            "nullable": true
          },
          "acceptedValuesJson": {
            "type": "string",
            "nullable": true
          },
          "isRegulatoryGate": {
            "type": "boolean"
          },
          "requiresVerification": {
            "type": "boolean"
          },
          "questionText": {
            "type": "string",
            "nullable": true
          },
          "expectedAnswer": {
            "type": "string",
            "nullable": true
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter: Bearer {your JWT token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}